src/main/java/jp/co/axa/apidemo/services/EmployeeService.java
package jp.co.axa.apidemo.services;

import jp.co.axa.apidemo.entities.Employee;

import java.util.List;

public interface EmployeeService {

    public List<Employee> retrieveEmployees();

    public Employee getEmployee(Long employeeId);

    public List<Employee> getEmployeesByName(String employeeName);

    public List<Employee> getEmployeesByDepartment(String department);

    public void saveEmployee(Employee employee);

    public void deleteEmployee(Long employeeId);

    public void updateEmployee(Employee employee);
}